# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateSnippetRequest:
    properties:
      content:
        description: 100KB max
        maxLength: 100000
        type: string
      label:
        maxLength: 255
        type: string
      shortcut:
        description: Short string without spaces
        maxLength: 50
        type: string
      tags:
        description: Max 20 tags, each max 50 chars
        items:
          type: string
        maxItems: 20
        type: array
    required:
    - content
    - label
    - shortcut
    type: object
  models.CreateUserRequest:
    properties:
      avatarUrl:
        maxLength: 500
        type: string
      email:
        maxLength: 255
        type: string
      fullName:
        maxLength: 255
        type: string
      password:
        description: Min 8 chars for security
        maxLength: 128
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.LoginRequest:
    properties:
      login:
        description: Can be username or email
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  models.LoginResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        description: Access token expiration in seconds
        type: integer
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  models.RefreshTokenResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        description: Access token expiration in seconds
        type: integer
    type: object
  models.Snippet:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      label:
        type: string
      shortcut:
        type: string
      tags:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  models.UpdateSnippetRequest:
    properties:
      content:
        type: string
      label:
        type: string
      shortcut:
        type: string
      tags:
        items:
          type: string
        type: array
      userId:
        description: UUID as string
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      avatarUrl:
        type: string
      email:
        type: string
      fullName:
        type: string
      password:
        description: Will be hashed
        type: string
      username:
        type: string
    type: object
  models.User:
    properties:
      avatarUrl:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        description: UUID as string
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Code snippets management API with authentication
  title: Snippy API
  version: "1.0"
paths:
  /auth/availability:
    get:
      description: Quickly verify whether a username or email can be used
      parameters:
      - description: Username to check
        in: query
        name: username
        type: string
      - description: Email to check
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check username/email availability
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with username/email and password
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the refresh token
      parameters:
      - description: Refresh token to revoke
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout
      tags:
      - auth
  /auth/logout-all:
    post:
      description: Revoke all refresh tokens for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout from all devices
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /snippets:
    get:
      description: Get all snippets belonging to the authenticated user
      parameters:
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Search in label
        in: query
        name: search
        type: string
      - description: Limit results (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's snippets
      tags:
      - snippets
    post:
      consumes:
      - application/json
      description: Create a new snippet for the authenticated user
      parameters:
      - description: Snippet data
        in: body
        name: snippet
        required: true
        schema:
          $ref: '#/definitions/models.CreateSnippetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Snippet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a snippet
      tags:
      - snippets
  /snippets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a snippet (owner only)
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a snippet
      tags:
      - snippets
    get:
      consumes:
      - application/json
      description: Get a single snippet by its ID
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Snippet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get snippet by ID
      tags:
      - snippets
    put:
      consumes:
      - application/json
      description: Update an existing snippet (owner only)
      parameters:
      - description: Snippet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: snippet
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSnippetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Snippet'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a snippet
      tags:
      - snippets
  /users:
    get:
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete user account (own account only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get a single user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user profile (own account only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/profile:
    get:
      description: Get the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the profile of the authenticated user
      parameters:
      - description: Update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"